requirements.txt
botocore==1.20.20
pandas==0.24.2
numpy==1.16.6
boto3==1.17.20
requests==2.18.4
urllib3==1.22
six==1.11.0
apispec==4.7.1
confluent_kafka==1.7.0
dbfread==2.0.7
falcon==3.0.1
falcon_apispec==0.4.0
falcon_cors==1.1.7
falcon_multipart==0.2.0
falcon_prometheus==0.1.0
falcon_swagger_ui==1.2.1
grequests==0.6.0
marshmallow==3.12.2
pycryptodome==3.10.1
pypika==0.48.5
python_bcrypt==0.3.2
python_dateutil==2.8.1
redis==3.5.3
uber_rides==0.6.0
xlrd==2.0.1
